coquitts/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
coqui-tts==0.27.2
soundfile==0.12.1
numpy==1.26.0
